chiselc: Move filter emitting to module
